update to 0.0.8 and tests in progress
