--- roshambo/code/.build
@@ -1,11 +1,0 @@
-{
-  "steps": {
-    "cargo": [
-      "build",
-      "--release",
-      "--target=wasm32-unknown-unknown",
-      "--target-dir=target"
-    ]
-  },
-  "artifact": "target/wasm32-unknown-unknown/release/roshambo.wasm"
-}
