--- roshambo/code/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "roshambo"
-version = "0.1.0"
-authors = ["ezra"]
-edition = "2018"
-
-[dependencies]
-serde = "1.0"
-serde_json = "1.0"
-serde_derive = "1.0"
-hdk = { git = "https://github.com/holochain/holochain-rust", tag = "v0.0.4-alpha" }
-holochain_wasm_utils = { git = "https://github.com/holochain/holochain-rust", tag = "v0.0.4-alpha" }
-holochain_core_types_derive = { git = "https://github.com/holochain/holochain-rust", tag = "v0.0.4-alpha" }
-multihash = "0.8.0"
-rand = "0.6.5"
-
-[lib]
-path = "src/lib.rs"
-crate-type = ["cdylib"]
