--- roshambo/zome.json
@@ -1,3 +1,0 @@
-{
-  "description": "The roshambo App"
-}
