+++ test/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "@holochain/holochain-nodejs": "0.4.2-alpha",
+    "@holochain/holochain-nodejs": "0.4.7-alpha",
     "json3": "*",
     "tape": "^4.9.1"
   }
