+++ zomes/roshambo/zome.json
@@ -1,0 +1,3 @@
+{
+  "description": "The roshambo App"
+}
